// src/serverApi/v3/models/copy-api-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CopyApiResponse
 */
export interface CopyApiResponse {
    /**
     * Id of copied element
     * @type {string}
     * @memberof CopyApiResponse
     */
    id?: string;
    /**
     * Title of copied element
     * @type {string}
     * @memberof CopyApiResponse
     */
    title?: string;
    /**
     * Type of copied element
     * @type {string}
     * @memberof CopyApiResponse
     */
    type: CopyApiResponseType;
    /**
     * Id of destination course
     * @type {string}
     * @memberof CopyApiResponse
     */
    destinationCourseId?: string;
    /**
     * Copy progress status of copied element
     * @type {string}
     * @memberof CopyApiResponse
     */
    status: CopyApiResponseStatus;
    /**
     * List of included sub elements with recursive type structure
     * @type {Array<CopyApiResponse>}
     * @memberof CopyApiResponse
     */
    elements?: Array<CopyApiResponse>;
}

/**
    * @export
    * @enum {string}
    */
export enum CopyApiResponseType {
    BOARD = 'BOARD',
    CONTENT = 'CONTENT',
    COURSE = 'COURSE',
    COURSEGROUP_GROUP = 'COURSEGROUP_GROUP',
    FILE = 'FILE',
    FILE_GROUP = 'FILE_GROUP',
    LEAF = 'LEAF',
    LESSON = 'LESSON',
    LESSON_CONTENT_ETHERPAD = 'LESSON_CONTENT_ETHERPAD',
    LESSON_CONTENT_GEOGEBRA = 'LESSON_CONTENT_GEOGEBRA',
    LESSON_CONTENT_GROUP = 'LESSON_CONTENT_GROUP',
    LESSON_CONTENT_LERNSTORE = 'LESSON_CONTENT_LERNSTORE',
    LESSON_CONTENT_NEXBOARD = 'LESSON_CONTENT_NEXBOARD',
    LESSON_CONTENT_TASK = 'LESSON_CONTENT_TASK',
    LESSON_CONTENT_TEXT = 'LESSON_CONTENT_TEXT',
    LERNSTORE_MATERIAL = 'LERNSTORE_MATERIAL',
    LERNSTORE_MATERIAL_GROUP = 'LERNSTORE_MATERIAL_GROUP',
    LTITOOL_GROUP = 'LTITOOL_GROUP',
    METADATA = 'METADATA',
    SUBMISSION_GROUP = 'SUBMISSION_GROUP',
    TASK = 'TASK',
    TASK_GROUP = 'TASK_GROUP',
    TIME_GROUP = 'TIME_GROUP',
    USER_GROUP = 'USER_GROUP'
}
/**
    * @export
    * @enum {string}
    */
export enum CopyApiResponseStatus {
    SUCCESS = 'success',
    FAILURE = 'failure',
    NOT_DOING = 'not-doing',
    NOT_IMPLEMENTED = 'not-implemented',
    PARTIAL = 'partial'
}